refs #7
classify response single or multiple
