+++ lib/rapa.rb
@@ -16,9 +16,11 @@ require "rapa/resources/base_resource"
 require "rapa/resources/browse_node_resource"
 require "rapa/resources/item_resource"
 require "rapa/responses/base_response"
-require "rapa/responses/list_items_response"
-require "rapa/responses/lookup_browse_node_response"
-require "rapa/responses/search_items_response"
+require "rapa/responses/multiple_resources_response"
+require "rapa/responses/multiple_resources_response/list_items_response"
+require "rapa/responses/multiple_resources_response/search_items_response"
+require "rapa/responses/single_resource_response"
+require "rapa/responses/single_resource_response/lookup_browse_node_response"
 require "rapa/signer"
 require "rapa/similar_product"
 require "rapa/url"
