+++ lib/rapa/responses/multiple_resources_response/search_items_response.rb
@@ -1,6 +1,6 @@
 module Rapa
   module Responses
-    class SearchItemsResponse < BaseResponse
+    class SearchItemsResponse < MultipleResourcesResponse
       # @return [Rapa::Error]
       def error
         if value = body.dig("ItemSearchResponse", "Items", "Request", "Errors", "Error")
