+++ lib/rapa/responses/single_resource_response.rb
@@ -1,0 +1,6 @@
+module Rapa
+  module Responses
+    class SingleResourceResponse < BaseResponse
+    end
+  end
+end
